Get rid of the uint32 for numWorkers, use zero-value for idlingCounter

+++ s5cmd.go
@@ -158,7 +158,7 @@ func main() {
 
 	wp := core.NewWorkerPool(ctx,
 		&core.WorkerPoolParams{
-			NumWorkers:     uint32(numWorkers),
+			NumWorkers:     numWorkers,
 			ChunkSizeBytes: multipartChunkSizeBytes,
 			Retries:        retries,
 		}, &s)
